styling and clean up

+++ src/app/components/post-list/post-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+
 import { Store, select } from '@ngrx/store';
 import { Observable } from 'rxjs';
 
@@ -7,7 +8,6 @@ import { PostItem } from '../../models/postItem.model';
 import * as fromPostActions from '../../store/actions/post.actions';
 import * as fromPostReducer from '../../store/reducers/post.reducer'
 import { AuthService } from '../../services/auth.service'
-import { ActivatedRoute, Router } from '@angular/router'
 
 @Component({
   selector: 'app-post-list',
@@ -19,10 +19,9 @@ export class PostListComponent implements OnInit {
   postItems$: Observable<Array<PostItem>>;
   loading$: Observable<Boolean>;
   error$: Observable<Error>
-  newPostItem: PostItem = { id: '', userid: '', title: '', body: '' }
   user: any;
 
-  constructor(private store: Store<AppState>, public auth: AuthService, private route: ActivatedRoute, private router: Router) { }
+  constructor(private store: Store<AppState>, public auth: AuthService) { }
 
   ngOnInit() {
     this.postItems$ = this.store.pipe(select(fromPostReducer.getPosts));
@@ -33,7 +32,9 @@ export class PostListComponent implements OnInit {
   }
 
   deletePost(id) {
-    this.store.dispatch(new fromPostActions.DeleteItemAction(id))
+    if (confirm("Are You Sure You want to Delete the Post?")) {
+      this.store.dispatch(new fromPostActions.DeleteItemAction(id))
+    }
   }
 
   editPost(id) {
